requirements.txt
scikit-learn
numpy
matplotlib
pandas
requests
python-dotenv
requests
wget